:sparkles: Use LLM to generate a theme/summary for each release (#9)
Needs to use the full log output instead of one-line. Feels like a paragraph
form is preferred to a bulleted list, since most of the rest of the release
doc is bullets. Section is called out as a summary or highlights.

Needed experimenting with different repos to get a flexible, good-enough
prompt.

Fix #9